OSX love: +Fixed broken context menus. Important for the Resources window for example, where right-clicking auto-save and auto-fill buttons for options was impossible on OSX! +Cycle action editor: added action learn, right-click in the right list view. Note: the editor is now exactly the same as on Windows OS. +Live Configs: added action learn (right-click in "Activation" and "Deactivation". Note: the Live Configs window is now exactly the same as on Windows OS. +Better tooltips
+Improved Region Playlist for new v4.23pre smooth seek options
+Fixed possible cue buss volume http://forum.cockos.com/showpost.php?p=949269&postcount=1378|issue|
+Fixed xcode 4 warnings

+++ SnM/SnM_ME.cpp
@@ -130,12 +130,12 @@ void MESetCCLanes(COMMAND_T* _ct)
 	{
 		// recall lanes
 		char laneSlot[MAX_CC_LANES_LEN], slot[32] = "";
-		_snprintf(slot, 32, "CC_LANES_SLOT%d", _ct->user + 1);
+		_snprintf(slot, 32, "CC_LANES_SLOT%d", (int)_ct->user + 1);
 		GetPrivateProfileString("MIDI_EDITOR", slot, "", laneSlot, MAX_CC_LANES_LEN, g_SNMIniFn.Get());
 
 		int i=0; 
 		while (laneSlot[i] && i < (MAX_CC_LANES_LEN-2)) // -2: see string termination
-		{ 
+		{
 			if (laneSlot[i] == '|')
 				laneSlot[i] = '\n';
 			i++;
@@ -193,7 +193,7 @@ void MESaveCCLanes(COMMAND_T* _ct)
 
 					// store lanes
 					char slot[32] = "";
-					_snprintf(slot, 32, "CC_LANES_SLOT%d", _ct->user + 1);
+					_snprintf(slot, 32, "CC_LANES_SLOT%d", (int)_ct->user + 1);
 					WritePrivateProfileString("MIDI_EDITOR", slot, laneSlot, g_SNMIniFn.Get());
 				}
 			}
